+++ src/entities/Row.ts
