+++ src/utils/EventManager.ts
